requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.8
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
python-dotenv==1.0.0
manim
pdfminer.six
elevenlabs
moviepy
black
pylint
pylatexenc
transformers
torch
tensorflow
langchain[llms]
python-dotenv
tiktoken
poe-api
chromadb
sentence_transformers
moviepy
imageio_ffmpeg